--- src/computations/model/solver/OutcomeSolver.java
@@ -1,8 +1,0 @@
-package computations.model.solver;
-
-import computations.model.DataRecord;
-
-public interface OutcomeSolver
-{
-	public int predictOutcome(DataRecord predict);
-}
