+++ src/computations/predictor/ml/ClockSpeed.java
@@ -5,10 +5,6 @@ class ClockSpeed
 	double time;
 	double speed;
 
-	public ClockSpeed()
-	{
-	}
-
 	@Override
 	public String toString()
 	{
