+++ src/computations/predictor/ml/model/solver/OutcomeSolver.java
@@ -1,0 +1,8 @@
+package computations.predictor.ml.model.solver;
+
+import computations.predictor.ml.model.DataRecord;
+
+public interface OutcomeSolver
+{
+	public int predictOutcome(DataRecord predict);
+}
