+++ src/computations/predictor/ml/solver/AllKnownMeasuresSolver.java
@@ -3,9 +3,9 @@ package computations.predictor.ml.solver;
 import java.util.ArrayList;
 import java.util.List;
 
-import computations.model.DataRecord;
-import computations.model.OutcomeStatistics;
+import computations.predictor.OutcomeStatistics;
 import computations.predictor.ml.PredictorML;
+import computations.predictor.ml.model.DataRecord;
 import logger.Logger;
 import servlets.SessionNotReadyException;
